type_checking: Added mutable type check
Now, the type checker will also bind expressions to their type.
In order to do that, we need a new mutable expression visitor.
The type checker has been modified to support a mutable expression
visitor. Because of this, the type checker is now using a mutable
statement visitor.

A typed trait allow to difference parts of the ast that are supposed to
have type to the others.